Refactor project according to class template
